Improve logger class with header, default filename and path, helper functions
The csv has now a header with column names. Also, there is a maximal number
of lines (600000) which corresponds to around 10 minutes of data recording.
The helper functions simplify the streaming of the data.

There is a default filename now, the current time (which is wrong by 1hour,
probably wrong timezone...) and the location where the file is created is
/tmp, which should be okay independent of the environment. The new bash script
copies all csv files from the remote host to the CITRIFIED/control/data directory.

There could be more checks if the filename and path are correct but this
is quite tiresome in CPP and we should just expect that the user know what
he's doing for now.Improve logger class with header, default filename and path, helper functions