Add NetFTRDTDriver for FT sensor and FT sensor class
The NetFTRDTDriver class is the same as in the net-ft-ros repository, except that
it's independent of ROS. There are a few TODOs there(new ones and copied ones), but
it works and that's what matters most at this point.

The ForceTorqueSensor class creates a NetFTRDTDriver to read the sensor and transforms
it into a StateRepresentation::CartesianWrench. Additionally, it computes the bias and
updates the actual readings with the bias.

The ft_sensor_test is just a check if everything works fine and shows how the FT sensor
could be used in combination with controllers.