Port RingDS trials to circle cut executable
* Move the lab day test changes from simple_ds_controller
to circular cut, since the simple ds example should
the point attractor demo.

For now, executables are a bit of a mixed bag between actual
runtime tasks and some kind of test. An upcoming step is
to create some simple unit tests so that the math of the
DS functions / controllers can be checked without having
to rewrite the task executable each time.