Move netutils header to franka_lwi_utils
* Move network/netutils.h to franka_lwi/franka_lwi_utils.h

* Change namespace to frankalwi::utils

* Refactor usages

The netutils header contained only franka_lwi related components,
and it also mixed a conversion util poseFromState() with the
configure() utility. By moving this header to franka_lwi include
directory and renaming to franka_lwi_utils.h, the structure
is tidier and the scope of the header supports both functions.