Add Ring DS (#26)
* Add RingDS definition

* Implement the RingDS function from the report
https://www.overleaf.com/read/ffqddknqnnvr

* Refinements to RingDS

* Make fixes to RingDS equations so that it functions better

* Tested on simple_ds_controller, gives good results at 4cm radius

This is not a clean commit, just a result of lab-day testing. Will
clean up before merging

* Refactor RingDS structure and document

* Revise class members and function names to have a nicer pattern.
Behaviour should be unchanged, though there are not yet unit tests
to verify this.

* Add doxygen headers to function implementations

* Port RingDS trials to circle cut executable

* Move the lab day test changes from simple_ds_controller
to circular cut, since the simple ds example should
the point attractor demo.

For now, executables are a bit of a mixed bag between actual
runtime tasks and some kind of test. An upcoming step is
to create some simple unit tests so that the math of the
DS functions / controllers can be checked without having
to rewrite the task executable each time.