TwistController
* Add TwistController

* Use external 'control_libraries::controllers' base controller
and patterns to replicate the old CartesianPoseControl

* Update mocap and simple point attractor to use new twist
controller and better state_rep practices

Early tests show identical behaviour to old class, just in a
much more compact fashion.

* Remove passiveDS and CartesianPoseController

I confirmed that the new TwistController using the Dissipative
controller in linear space works the same as the passiveDS
controller. Therefore I am removing all external dependecies
to passive-ds, including the lib directory. The CartesianPoseCtrl
is now redundant, so it has also been removed. The remaining
reference to CartesianPoseController in circular_cut has
been replaced with TwistController.

The executables have been checked on the robot, and the
build script passes as well.