First order hold for derivative boundaries
* Derivatives are calculated as centered finite differences, which
means that the first and last sample have no calculation.
At first they were left as zeros (zero-order hold) in the
boundaries, but this caused some jumps when plotting
the force derivative (particularly when the force
appeared to have a constant derivative). Now, it's
using first-order hold, so that the derivative is held
constant at the boundaries.