ESNWrapper for incision trials (#49)
* ESNWrapper for incision trials

* Add revised ESNWrapper to handle real-time running task
of classification. Add timestamped samples, and then
run a background thread to automatically latch a buffered
sample window, calculate force derivatives and return
an esn prediction

* Integrate ESNWrapper into incision_trials

* Add subfield and getTime methods to jsonlogger for
convenience

* Add inputDimensions() getter to ESN class for convenience

* Make a basic test for ESNWrapper (without threading)

* Fix control lib version change issue in franka_lwi test
so that the test target can build

* Log ESN output

* Add ESN output to json logger

* Add depth sample to input buffer, and only addSamples when
it's actually needed (in insertion phase)

* Revise data format spec for ESN

* Various fixes

* Add yaml file as string to metadata

* Set logger file name properly

* Add depth input to metadata

* Add esn skip logic to downsample frequency

* Log static configurations of optitrack

* Fix conversion bug from data buffers to STl types for json

* Fix esn input format

* Get data buffer columns instead of rows

* First order hold for derivative boundaries

* Derivatives are calculated as centered finite differences, which
means that the first and last sample have no calculation.
At first they were left as zeros (zero-order hold) in the
boundaries, but this caused some jumps when plotting
the force derivative (particularly when the force
appeared to have a constant derivative). Now, it's
using first-order hold, so that the derivative is held
constant at the boundaries.