Fix cut prober height estimate
* Revise estimateHeightInTask function to stabilise
interpolation

* Improve unit test for height estimate

Before, the height was linearly interpolated as the
weighted average between the two closest sample points.
As the sample points are not spaced perfectly
evenly, this can cause the "two closest" sample points
to not necessarily be the bounding points on the cut line.

Instead, the assumption of chronologically ordered surface
samples allows the other bounding sample to be the one
directly before or after the minimum distance sample.
The dot product is used to find which is appropriate.